+++ react/BankInvoiceSection.tsx
@@ -29,8 +29,9 @@ const BankInvoiceSection: FC = () => {
 
   const { url, paymentSystemName, barCodeNumber } = paymentInfo
   const isURLValid = url && !url.match(/(\*.\*.)+\*\w\*/g)
+  const hideBankInvoice = !isURLValid && !barCodeNumber
 
-  if (!isURLValid && !barCodeNumber) {
+  if (hideBankInvoice) {
     return null
   }
 
@@ -58,7 +59,7 @@ const BankInvoiceSection: FC = () => {
         className={`${handles.barCodeContainer} flex-l justify-between items-center mt6`}
       >
         {barCodeNumber && <BarCode barCodeNumber={barCodeNumber} />}
-        {isURLValid && (
+        {parsedUrl && (
           <div
             className={`${handles.printButtonWrapper} mt5 mt0-l ${
               barCodeNumber ? 'ml5-l' : ''